refactor: move wildcard expansion loop
